+++ src/main/java/com/epam/newsmanagement/model/persistence/interfaces/TagDAO.java
@@ -7,8 +7,8 @@ import com.epam.newsmanagement.model.persistence.exception.DAOException;
 import java.util.List;
 
 public interface TagDAO extends GenericDAO<Tag> {
-    List<Integer> insert(List<Tag> tags) throws DAOException;
-    Tag findById(int id);
-    List<Tag> findByNewsId(int newsId);
+    List<Long> insert(List<Tag> tags) throws DAOException;
+    Tag findById(long id);
+    List<Tag> findByNewsId(long newsId);
     Tag findByName(String name);
 }
